Refactor logger
