+++ lib/helpers/logger/index.js
@@ -1,5 +1,5 @@
 const bunyan = require('bunyan');
-const PrettyStdout = require('./logger-pretty-stdout');
+const PrettyStdout = require('./pretty-stdout');
 const traverse = require('traverse');
 
 const prettyStdOut = new PrettyStdout();
