+++ lib/helpers/logger/pretty-stdout.js
