+++ test/helpers/logger.spec.js
@@ -1,0 +1,9 @@
+const loggerHelper = require('../../lib/helpers/logger');
+
+describe('helpers/logger', () => {
+  describe('configSerializer(config)', () => {
+    it('something', async () => {
+      // nothing
+    });
+  });
+});
